monitoring/requirements.txt
docker=5.0.3
pandas=1.4.2
pytest=7.1.2
python-dotenv=0.20.0
argparse=1.1
psutils=5.9.0
kaleido=0.2.1